feat/todo-app#10: create entity for user modify repo name for refresh token repository
