+++ backend/src/modules/users/entites/users.entity.ts
@@ -1,0 +1,8 @@
+export type TUser = { 
+    id: string 
+    email: string 
+    password: string 
+    name: string 
+    createdAt: Date 
+    updatedAt: Date 
+};
