+++ backend/src/modules/users/repositories/user.tokens.ts
@@ -1,0 +1,1 @@
+export const USER_REPOSITORY = 'USER_REPOSITORY';
