+++ backend/src/modules/users/users.module.ts
@@ -4,6 +4,7 @@ import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { UserRepository } from './repositories/user.repository';
 import { PrismaUserRepository } from './repositories/prisma-user.repository';
+import { USER_REPOSITORY } from './repositories/user.tokens';
 
 @Module({
   imports: [PrismaModule],
@@ -11,10 +12,10 @@ import { PrismaUserRepository } from './repositories/prisma-user.repository';
   providers: [
     UsersService,
     {
-      provide: UserRepository,
+      provide: USER_REPOSITORY,
       useClass: PrismaUserRepository,
     },
   ],
-  exports: [UserRepository],
+  exports: [USER_REPOSITORY],
 })
 export class UsersModule { }
